Move config_table_structure to keep correct formats
